requirements.txt
telebot~=0.0.5
requests~=2.31.0
python-dotenv~=1.0.1
SQLAlchemy~=2.0.31
psycopg2-binary~=2.9.9